requirements.txt
dataset==0.8.0
discord.py==0.16.8
